+++ js/loaders/schemas/loadSchema3Event.js
@@ -1,21 +1,18 @@
-import { linkTypes } from "../lib/constants/linkTypes.js";
-import { getSupportedEDM4hepTypes } from "../state/globals.js";
+import { linkTypes } from "../../lib/constants/linkTypes.js";
+import { getSupportedEDM4hepTypes } from "../../state/globals.js";
 import { loadPlainObject } from "./loadObjects.js";
 
-export function handleSchema1Event(eventData) {
+export function handleSchema3Event(eventData) {
   const objects = {
     datatypes: {},
     associations: {},
   };
 
-  const supportedEDM4hepTypes = getSupportedEDM4hepTypes("1");
+  const supportedEDM4hepTypes = getSupportedEDM4hepTypes("3");
 
   // Select only Datatype collections
   const supportedDataTypes = Object.keys(supportedEDM4hepTypes).filter(
     (type) => {
-      if (type.includes("Association")) {
-        return false;
-      }
       if (type.includes("Link")) {
         return false;
       }
@@ -26,7 +23,7 @@ export function handleSchema1Event(eventData) {
 
   // Select only Link collections
   const supportedAssociations = Object.keys(supportedEDM4hepTypes).filter(
-    (object) => object.includes("Association"),
+    (object) => object.includes("Link"),
   );
 
   supportedDataTypes.forEach((typeName) => {
@@ -69,7 +66,7 @@ export function handleSchema1Event(eventData) {
         supportedDataType,
         collectionId,
         collName,
-        "1",
+        "3",
       );
       objects.datatypes[supportedDataType].collection.push(...objectCollection);
     }
@@ -100,7 +97,8 @@ export function handleSchema1Event(eventData) {
           if (objects.datatypes?.[type] === undefined) continue;
           const oneToOneRelationData = element.collection
             .map((object) => object[name])
-            .filter((object) => object !== undefined);
+            .filter((object) => object !== undefined)
+            .map((object) => object[0]);
 
           if (oneToOneRelationData.length === 0) continue;
 
@@ -184,8 +182,8 @@ export function handleSchema1Event(eventData) {
   // Currently, all associations are one-to-one
   for (const association of supportedAssociations) {
     Object.values(eventData).forEach((element) => {
-      const collectionName = `${association}Collection`;
-      if (element.collType === collectionName) {
+      // const collectionName = `${association}Collection`;
+      if (element.collType === association) {
         const collection = element.collection;
         if (collection.length === 0) return;
 
@@ -215,17 +213,25 @@ export function handleSchema1Event(eventData) {
           const fromObject = fromCollection[associationElement[fromName].index];
           const toObject = toCollection[associationElement[toName].index];
 
+          /** linkTypes[association] receives:
+           * - podio::LinkCollection<edm4hep::ReconstructedParticle,edm4hep::MCParticle>
+           * - @todo not supported - podio::LinkCollection<edm4hep::Cluster,edm4hep::MCParticle>
+           * - @todo not supported - podio::LinkCollection<edm4hep::Track,edm4hep::MCParticle>
+           */
           const linkType = linkTypes[association];
-          const link = new linkType(
-            fromObject,
-            toObject,
-            associationElement.weight,
-          );
-          objects.associations[association].push(link);
-          fromObject.associations = {};
-          fromObject.associations[association] = link;
-          toObject.associations = {};
-          toObject.associations[association] = link;
+
+          if (linkType) {
+            const link = new linkType(
+              fromObject,
+              toObject,
+              associationElement.weight,
+            );
+            objects.associations[association].push(link);
+            fromObject.associations = {};
+            fromObject.associations[association] = link;
+            toObject.associations = {};
+            toObject.associations[association] = link;
+          }
         }
       }
     });
