+++ js/loaders/schemas/loadSchema4Event.js
@@ -1,5 +1,5 @@
-import { linkTypes } from "../viz/links/linkTypes.js";
-import { getSupportedEDM4hepTypes } from "../state/globals.js";
+import { linkTypes } from "../../viz/links/linkTypes.js";
+import { getSupportedEDM4hepTypes } from "../../state/globals.js";
 import { loadPlainObject } from "./loadObjects.js";
 
 export function handleSchema4Event(eventData) {
